requirements.txt
jira
openai==0.28
pyperclip
python-dotenv
pytz
requests
